requirement.txt
Django==3.2
Pillow==8.2.0
djangorestframework==3.12.4
django-rest-knox==4.1.0
django-filter-2.4.0
django-cors-headers-3.7.0